+++ CHANGELOG.md
@@ -3,6 +3,9 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Changed
+* `flexberry-menu` component:
+    * By default handles click and touchstart events.
 
 ## [2.0.0-beta.5] - 2018-11-16
 ### Changed